+++ src/jsonp.js
@@ -9,7 +9,6 @@ const FALSE = false;
 let removeScript = (script) => {
     script.onerror = NULL;
     script.parentNode.removeChild(script);
-    script = NULL;
 };
 let head = NULL;
 let insertScript = (url, options) => {