+++ test-src/storage.spec.js
@@ -11,13 +11,13 @@ describe('storage', function () {
     let context;
 
     beforeEach('reset', function () {
-        context = nattyFetch.context('storage', {
+        context = nattyFetch.context({
             urlPrefix: host,
             mock: false,
         });
     });
 
-    it('query string is same', function (done) {
+    it('query string is same: localStorage', function (done) {
         let requestTime = 0;
         context.create('user', {
             getPhone: {
@@ -29,6 +29,48 @@ describe('storage', function () {
                 },
                 storage: {
                     type: 'localStorage',
+                    tag: 'v1.0',
+                    key: 'test-query-string'
+                }
+            }
+        });
+
+        // 第一次请求走网络
+        context.api.user.getPhone({
+            b:1,
+            a:1
+        }).then(function (r) {
+            // 第二次请求走storage
+            return context.api.user.getPhone({
+                a:1,
+                b:1
+            });
+        }).then(function (data) {
+            try {
+                expect(data.id).to.be(1);
+                expect(requestTime).to.be(1);
+                done();
+            } catch (e) {
+                done(e);
+            }
+            // 特别注意 expect不管是否成功, 都要销毁storage, 避免下次刷新后的测试
+            context.api.user.getPhone.storage.destroy();
+
+        }).catch();
+    });
+
+    it('query string is same: sessionStorage', function (done) {
+        let requestTime = 0;
+        context.create('user', {
+            getPhone: {
+                url: host + 'api/return-success',
+                willFetch: function (vars, config, from) {
+                    if (from === 'remote') {
+                        requestTime++;
+                    }
+                },
+                storage: {
+                    type: 'sessionStorage',
                     tag: 'v1.0'
                 }
             }
@@ -54,6 +96,46 @@ describe('storage', function () {
             }
             // 特别注意 expect不管是否成功, 都要销毁storage, 避免下次刷新后的测试
             context.api.user.getPhone.storage.destroy();
+
+        }).catch();
+    });
+
+    //
+    it('query string is same: variable', function (done) {
+        let requestTime = 0;
+        context.create('user', {
+            getPhone: {
+                url: host + 'api/return-success',
+                willFetch: function (vars, config, from) {
+                    if (from === 'remote') {
+                        requestTime++;
+                    }
+                },
+                storage: {
+                    type: 'variable',
+                    tag: 'v1.0'
+                }
+            }
+        });
+
+        // 第一次请求走网络
+        context.api.user.getPhone({
+            b:1,
+            a:1
+        }).then(function (r) {
+            // 第二次请求走storage
+            return context.api.user.getPhone({
+                a:1,
+                b:1
+            });
+        }).then(function (data) {
+            try {
+                expect(data.id).to.be(1);
+                expect(requestTime).to.be(1);
+                done();
+            } catch (e) {
+                done(e);
+            }
         }).catch();
     });
 
@@ -70,7 +152,8 @@ describe('storage', function () {
                 },
                 storage: {
                     type: 'localStorage',
-                    tag: 'v1.0'
+                    tag: 'v1.0',
+                    key: 'test-jsonp-with-storage'
                 }
             }
         });